reducing code MockContext
